akena/03-agents/agent-oferta/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
anthropic>=0.40.0
python-multipart>=0.0.9
pypdf>=4.0.0
python-docx>=1.1.0
